✨ Add URL-based card anchoring with ?card= deep links
Add support for navigating directly to a specific dashboard card via
URL query parameter (e.g., /?card=cluster_health). The card scrolls
into view and flashes a purple highlight ring.

- Extract scrollToCard utility from SearchDropdown into shared lib
- Extend useDeepLink hook to handle ?card= param with auto-cleanup
- Skip last-route redirect when deep link params are present
- Add "Copy Link" to every card's dropdown menu (CardWrapper)

Completes the half-built feature where MiniDashboard.tsx already
generated ?card= URLs but nothing consumed them.
