+++ web/src/components/layout/navbar/SearchDropdown.tsx
@@ -19,6 +19,7 @@ import {
 } from 'lucide-react'
 import { useSearchIndex, CATEGORY_ORDER, type SearchCategory, type SearchItem } from '../../../hooks/useSearchIndex'
 import { useMissions } from '../../../hooks/useMissions'
+import { scrollToCard } from '../../../lib/scrollToCard'
 
 const CATEGORY_CONFIG: Record<SearchCategory, { label: string; icon: typeof Server }> = {
   page: { label: 'Pages', icon: LayoutDashboard },
@@ -34,36 +35,6 @@ const CATEGORY_CONFIG: Record<SearchCategory, { label: string; icon: typeof Serv
   dashboard: { label: 'Dashboards', icon: LayoutDashboard },
   helm: { label: 'Helm Releases', icon: Package },
   node: { label: 'Nodes', icon: HardDrive },
-}
-
-const SCROLL_HIGHLIGHT_MS = 2000
-const SCROLL_POLL_INTERVAL_MS = 100
-const SCROLL_POLL_MAX_MS = 3000
-
-/**
- * After navigation, poll for a card element by data-card-type and scroll it into view
- * with a brief highlight ring.
- */
-function scrollToCard(cardType: string) {
-  const startTime = Date.now()
-
-  function poll() {
-    const el = document.querySelector(`[data-card-type="${cardType}"]`)
-    if (el) {
-      el.scrollIntoView({ behavior: 'smooth', block: 'center' })
-      el.classList.add('ring-2', 'ring-purple-500', 'ring-offset-2', 'ring-offset-background')
-      setTimeout(() => {
-        el.classList.remove('ring-2', 'ring-purple-500', 'ring-offset-2', 'ring-offset-background')
-      }, SCROLL_HIGHLIGHT_MS)
-      return
-    }
-    if (Date.now() - startTime < SCROLL_POLL_MAX_MS) {
-      setTimeout(poll, SCROLL_POLL_INTERVAL_MS)
-    }
-  }
-
-  // Start polling after a frame so React can render the new route
-  requestAnimationFrame(() => setTimeout(poll, SCROLL_POLL_INTERVAL_MS))
 }
 
 export function SearchDropdown() {
